# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Tangent Blog OpenApi Demo Documentation",
        "description": "Tangent Blog Swagger OpenApi description",
        "contact": {
            "email": "sipho411@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://tangent_blog.test",
            "description": "Demo Tangent Blog API Server"
        }
    ],
    "paths": {
...
        "/api/comments": {
            "get": {
                "tags": [
                    "comments"
                ],
                "summary": "Get all comments by auth user.",
                "description": "Get all comments by auth user.",
                "operationId": "commentList",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "App\\Http\\Resources\\Api\\V1\\CommentResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/comment/store": {
            "post": {
                "tags": [
                    "comments"
                ],
                "summary": "Add comment as Auth user..",
                "description": "Add comment as auth user.",
                "operationId": "commentStore",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "App\\Http\\Resources\\Api\\V1\\CommentResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/comment/{id}/show": {
            "get": {
                "tags": [
                    "comments"
                ],
                "summary": "Get comment details.",
                "description": "Get comment details.",
                "operationId": "commentDetails",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "App\\Http\\Resources\\Api\\V1\\CommentResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/comment/{id}/update": {
            "post": {
                "tags": [
                    "comments"
                ],
                "summary": "Update comment",
                "description": "Update comment",
                "operationId": "commentUpdate",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "App\\Http\\Resources\\Api\\V1\\CommentResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/comment/{id}/destroy": {
            "post": {
                "tags": [
                    "comments"
                ],
                "summary": "Delete comment",
                "description": "Delete Comment..",
                "operationId": "commentDestroy",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "App\\Http\\Resources\\Api\\V1\\CommentResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
...
    },
...
}